+++ packages/adapters-file/src/index.ts
@@ -1,1 +1,2 @@
-export * from './FileAdapter.js'; 
+export * from './FileAdapter.js';
+export * from './FilePromptRepository';