+++ packages/core/src/index.ts
@@ -2,9 +2,27 @@
 export * from './entities/Prompt';
 export * from './entities/TemplateVariable';
 export * from './entities/PromptSequence';
-export * from './ports/IPromptApplication';
+export * from './value-objects/PromptId';
 export * from './ports/IPromptRepository';
+export * from './ports/IPromptApplication';
 export * from './ports/ITemplatingEngine';
 export * from './ports/IEventPublisher';
 export * from './ports/ISecurityValidator';
+export * from './use-cases/addPrompt';
+export * from './use-cases/getPromptById';
+export * from './use-cases/listPrompts';
+export * from './use-cases/updatePrompt';
+export * from './use-cases/deletePrompt';
+export type { Prompt } from './entities/Prompt';
+export type { PromptSequence } from './entities/PromptSequence';
+export type { WorkflowExecutionState, ListPromptsOptions } from './interfaces';
+export type { IPromptRepository } from './ports/IPromptRepository';
+export type { ISequenceRepository } from './interfaces';
+export type { IPromptApplication } from './ports/IPromptApplication';
+export type { PromptId } from './value-objects/PromptId';
+export { addPrompt } from './use-cases/addPrompt';
+export { getPromptById } from './use-cases/getPromptById';
+export { listPrompts } from './use-cases/listPrompts';
+export { updatePrompt } from './use-cases/updatePrompt';
+export { deletePrompt } from './use-cases/deletePrompt';
 // ...add more as you extract